Separate test that require network access in test/data/test_field.py
